+++ R/gghistostats.R
@@ -16,11 +16,10 @@
 #'   provides redundant information in light of y-axis labels.
 #' @param low.color,high.color Colors for low and high ends of the gradient.
 #'   Defaults are colorblind-friendly.
-#' @param normal.curve Logical decides whether to super-impose a normal curve
-#'   using `stats::dnorm(mean(x), sd(x))`. Default is `FALSE`.
-#' @param normal.curve.color,normal.curve.linetype,normal.curve.size If
-#'   `normal.curve = TRUE`, then these arguments can be used to modify color
-#'   (Default: `"black"`), size (default: `1.0`), linetype (default: `"solid"`).
+#' @param normal.curve A logical value that decides whether to super-impose a
+#'   normal curve using `stats::dnorm(mean(x), sd(x))`. Default is `FALSE`.
+#' @param normal.curve.args A list of additional aesthetic arguments to be
+#'   passed to the normal curve.
 #' @param bar.fill If `fill.gradient = FALSE`, then `bar.fill` decides which
 #'   color will uniformly fill all the bars in the histogram (Default:
 #'   `"grey50"`).
@@ -38,12 +37,11 @@
 #'
 #' @import ggplot2
 #'
-#' @importFrom dplyr select bind_rows summarize mutate mutate_at mutate_if
+#' @importFrom dplyr select summarize mutate
 #' @importFrom dplyr group_by n arrange
 #' @importFrom rlang enquo as_name !!
 #' @importFrom scales percent percent_format
 #' @importFrom stats dnorm
-#' @importFrom crayon green blue yellow red
 #' @importFrom statsExpressions expr_t_onesample bf_ttest
 #'
 #' @references
@@ -57,7 +55,7 @@
 #'   data = ToothGrowth,
 #'   x = len,
 #'   xlab = "Tooth length",
-#'   centrality.para = "median"
+#'   centrality.parameter = "median"
 #' )
 #'
 #' # a detailed function call
@@ -103,22 +101,16 @@ gghistostats <- function(data,
                          high.color = "#D55E00",
                          bar.fill = "grey50",
                          results.subtitle = TRUE,
-                         centrality.para = "mean",
-                         centrality.color = "blue",
-                         centrality.size = 1.0,
-                         centrality.linetype = "dashed",
-                         centrality.line.labeller = TRUE,
-                         centrality.k = 2,
-                         test.value.line = FALSE,
-                         test.value.color = "black",
-                         test.value.size = 1.0,
-                         test.value.linetype = "dashed",
-                         test.line.labeller = TRUE,
                          test.k = 0,
+                         test.value.line = FALSE,
+                         test.value.line.args = list(size = 1),
+                         test.value.label.args = list(),
+                         centrality.parameter = "mean",
+                         centrality.k = 2,
+                         centrality.line.args = list(color = "blue", size = 1),
+                         centrality.label.args = list(color = "blue"),
                          normal.curve = FALSE,
-                         normal.curve.color = "black",
-                         normal.curve.linetype = "solid",
-                         normal.curve.size = 1.0,
+                         normal.curve.args = list(),
                          ggplot.component = NULL,
                          output = "plot",
                          messages = TRUE,
@@ -182,8 +174,12 @@ gghistostats <- function(data,
   }
 
   # quit early if only subtitle is needed
-  if (output == "subtitle") {
-    return(subtitle)
+  if (output %in% c("subtitle", "caption")) {
+    return(switch(
+      EXPR = output,
+      "subtitle" = subtitle,
+      "caption" = caption
+    ))
   }
 
   # ============================= plot ====================================
@@ -317,13 +313,12 @@ gghistostats <- function(data,
 
     # adding curve to the plot
     plot <- plot +
-      ggplot2::stat_function(
+      rlang::exec(
+        .f = ggplot2::stat_function,
         fun = .f_stat,
-        linetype = normal.curve.linetype,
-        color = normal.curve.color,
-        size = normal.curve.size,
         na.rm = TRUE,
-        args = args
+        args = args,
+        !!!normal.curve.args
       )
   }
 
@@ -348,30 +343,24 @@ gghistostats <- function(data,
   # ====================== centrality line and label ========================
 
   # computing statistics needed for displaying labels
-  y_label_pos <- median(
-    x = ggplot2::layer_scales(plot)$y$range$range,
-    na.rm = TRUE
-  )
+  y_label_pos <- median(ggplot2::layer_scales(plot)$y$range$range, na.rm = TRUE)
 
   # using custom function for adding labels
-  plot <- histo_labeller(
-    plot = plot,
-    x = x_vec,
-    y.label.position = y_label_pos,
-    centrality.para = centrality.para,
-    centrality.color = centrality.color,
-    centrality.size = centrality.size,
-    centrality.linetype = centrality.linetype,
-    centrality.line.labeller = centrality.line.labeller,
-    centrality.k = centrality.k,
-    test.value = test.value,
-    test.value.line = test.value.line,
-    test.value.color = test.value.color,
-    test.value.size = test.value.size,
-    test.value.linetype = test.value.linetype,
-    test.line.labeller = test.line.labeller,
-    test.k = test.k
-  )
+  plot <-
+    histo_labeller(
+      plot = plot,
+      x = x_vec,
+      y.label.position = y_label_pos,
+      test.value = test.value,
+      test.k = test.k,
+      test.value.line = test.value.line,
+      test.value.line.args = test.value.line.args,
+      test.value.label.args = test.value.label.args,
+      centrality.parameter = centrality.parameter,
+      centrality.k = centrality.k,
+      centrality.line.args = centrality.line.args,
+      centrality.label.args = centrality.label.args
+    )
 
   # if no color fill gradient is used, then remove the legend
   if (isFALSE(fill.gradient)) plot <- plot + ggplot2::theme(legend.position = "none")
@@ -386,7 +375,7 @@ gghistostats <- function(data,
 
   # display normality test result as a message
   if (isTRUE(messages)) {
-    normality_message(
+    ipmisc::normality_message(
       x = x_vec,
       lab = xlab,
       k = k,
@@ -395,11 +384,5 @@ gghistostats <- function(data,
   }
 
   # return the final plot
-  return(switch(
-    EXPR = output,
-    "plot" = plot,
-    "subtitle" = subtitle,
-    "caption" = caption,
-    plot
-  ))
+  return(plot)
 }