+++ R/reexports.R
@@ -1,17 +1,17 @@
 
-# ----------------------- groupedstats -----------------------------------
+# ----------------------- ipmisc -----------------------------------
 
 #' @export
-#' @importFrom groupedstats specify_decimal_p
-groupedstats::specify_decimal_p
+#' @importFrom ipmisc specify_decimal_p
+ipmisc::specify_decimal_p
 
 #' @export
-#' @importFrom groupedstats signif_column
-groupedstats::signif_column
+#' @importFrom ipmisc signif_column
+ipmisc::signif_column
 
 #' @export
-#' @importFrom groupedstats set_cwd
-groupedstats::set_cwd
+#' @importFrom ipmisc set_cwd
+ipmisc::set_cwd
 
 # ----------------------- pairwiseComparisons --------------------------------
 
@@ -26,20 +26,20 @@ pairwiseComparisons::pairwise_comparisons
 # ----------------------- tidyverse -------------------------------------
 
 #' @export
-#' @importFrom broomExtra "%>%"
-broomExtra::`%>%`
+#' @importFrom ipmisc "%>%"
+ipmisc::`%>%`
 
 #' @export
-#' @importFrom broomExtra "%<>%"
-broomExtra::`%<>%`
+#' @importFrom ipmisc "%<>%"
+ipmisc::`%<>%`
 
 #' @export
-#' @importFrom broomExtra "%$%"
-broomExtra::`%$%`
+#' @importFrom ipmisc "%$%"
+ipmisc::`%$%`
 
 #' @export
-#' @importFrom statsExpressions "%<-%"
-statsExpressions::`%<-%`
+#' @importFrom ipmisc "%<-%"
+ipmisc::`%<-%`
 
 
 #' @export
@@ -57,3 +57,13 @@ tibble::tibble
 #' @export
 #' @importFrom tibble as_tibble
 tibble::as_tibble
+
+# ----------------------- messages -------------------------------------
+
+#' @export
+#' @importFrom ipmisc bartlett_message
+ipmisc::bartlett_message
+
+#' @export
+#' @importFrom ipmisc normality_message
+ipmisc::normality_message